update platform auth available spec
