--- app/javascript/packages/webauthn/is-public-key-credential-supported.ts
@@ -1,7 +1,0 @@
-async function isPublicKeyCredentialSupported(): Promise<boolean> {
-  const isUserVerifyingPlatformAuthenticatorAvailable =
-    await window.PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable();
-  return window.PublicKeyCredential && isUserVerifyingPlatformAuthenticatorAvailable;
-}
-
-export default isPublicKeyCredentialSupported;
