+++ app/javascript/packages/webauthn/is-webauthn-platform-authenticator-available.ts
@@ -1,0 +1,6 @@
+export type isWebauthnPlatformAvailable = () => Promise<boolean>;
+
+const isWebauthnPlatformAuthenticatorAvailable: isWebauthnPlatformAvailable = async () =>
+  !!(await window.PublicKeyCredential?.isUserVerifyingPlatformAuthenticatorAvailable());
+
+export default isWebauthnPlatformAuthenticatorAvailable;
